Fixed bugs and added features:
fixed bug: Only up to 2 additional key binding could be saved
added feature: in controls option can now only add new alternative
keybinding if main and other keybinding are bound
added feature: fancy toast if can not add new alternative keybinding
added feature: adding new alternative keybinding automatically sets the
new keybinding button active for key listening
fixed bug: reset keybinding and its alternatives tooltip now only shows
when it is clickable
added feature: pressing esc (unbinding) a key deletes that entry
fixed bug: game crashes when multiple alternatives keybindings have same
alternative id
fixed bug: keybindings are not saved when alternative keybindings get
deleted

code refactoring

+++ src/main/java/de/siphalor/nmuk/impl/IKeyBinding.java
@@ -17,16 +17,17 @@
 
 package de.siphalor.nmuk.impl;
 
-import net.minecraft.client.option.KeyBinding;
+import java.util.List;
+
 import org.jetbrains.annotations.ApiStatus;
 
-import java.util.List;
+import net.minecraft.client.option.KeyBinding;
 
 @ApiStatus.Internal
 public interface IKeyBinding {
-	short nmuk_getNextChildId();
+	int nmuk_getNextChildId();
 
-	void nmuk_setNextChildId(short nextChildId);
+	void nmuk_setNextChildId(int nextChildId);
 
 	boolean nmuk_isAlternative();
